resnet/requirements.txt
torch
torchvision
numpy<2
PyYAML
mlflow
Pillow
